Move `ReadableThrowable` to `:platform:autos`
