+++ core/mastodon/src/main/java/br/com/orcinus/orca/core/mastodon/feed/profile/MastodonProfileProvider.kt
@@ -19,7 +19,7 @@ import android.content.Context
 import br.com.orcinus.orca.core.feed.profile.Profile
 import br.com.orcinus.orca.core.feed.profile.ProfileProvider
 import br.com.orcinus.orca.core.mastodon.R
-import br.com.orcinus.orca.core.mastodon.i18n.ReadableThrowable
+import br.com.orcinus.orca.platform.autos.i18n.ReadableThrowable
 import br.com.orcinus.orca.platform.cache.Cache
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
