+++ platform/autos/src/test/res/values-pt-rBR/strings.xml
