+++ platform/autos/src/test/res/values/strings.xml
